+++ src/test/java/Dao/OrderDetailsDaoImplTest.java
@@ -49,7 +49,7 @@ public class OrderDetailsDaoImplTest {
     @Test
     public  void getOrderDetailsByMIdAndOIdTest(){
         int m_id =10001;
-        int o_id = 1;
+        long o_id = 1;
         List<OrderDetails> orderDetails = orderDetailsDao.getOrderDetailsByMIdAndOId(m_id,o_id);
         for(OrderDetails orderDetails1 : orderDetails){
             System.out.println(orderDetails1);
